requirements.txt
torch>=1.9.0
transformers>=4.15.0
pandas>=1.3.0
numpy>=1.21.0
Pillow>=8.0.0
pytesseract>=0.3.8
librosa>=0.8.1
scikit-learn>=0.24.2
faiss-cpu>=1.7.0
python-docx>=0.8.11
PyPDF2>=2.0.0
openpyxl>=3.0.7
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
soundfile>=0.10.3
sentence-transformers>=2.2.0 